Added patches for updating the script for the Minish font
This patch adjusts some of the credits being misaligned when using the less wider Minish font.

It also removes a comma in item "Kind Kiyo" that would blend with the bottom text.

It also replaces the hyphen in sky-high for "sky high" in item "RC Doctor". Just a style decision, mainly because there's already a hyphen for 'remote-controlled'.